// src/modules/invoice/usecase/find/find-invoice-dto.ts
export interface FindInvoiceUseCaseInputDTO {
    id: string
}

export interface FindInvoiceUseCaseOutputDTO {
    id: string
    name: string
    document: string
    address: {
        street: string
        number: string
        complement: string
        city: string
        state: string
        zipCode: string
    }
    items: {
        id: string
        name: string
        price: number
    }[]
    total: number
    createdAt: Date
}